detection/ourmodel/requirements.txt
torch>=1.7.1
torchvision
numpy
opencv-python
Pillow
fastapi
uvicorn
pydantic
timm
einops
scikit-image
scipy
matplotlib
tqdm
pandas
albumentations
efficientnet_pytorch
transformers